% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{add_emp_cols}
\alias{add_emp_cols}
\title{Add new columns as with NAs if not existed}
\usage{
add_emp_cols(df, vars)
}
\arguments{
\item{df}{Data frame}

\item{vars}{Vars to include or add.
`vars` not existed in the data frame will be added with NAs.
Existing `vars` are kept.}
}
\value{
A new Data Frame with all `vars`
}
\description{
Add new columns as with NAs if not existed
}
