% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_num_range}
\alias{check_num_range}
\title{Check the range of a list of numeric variables}
\usage{
check_num_range(df, vars)
}
\arguments{
\item{df}{the data frame to check}

\item{vars}{numeric variables}
}
\value{
return `vars` invisibly
}
\description{
Check the range of a list of numeric variables
}
\examples{
\dontrun{check_range(convert_choiceDF(choice_df), num_vars)}
}
