% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_range}
\alias{check_range}
\title{Check the range of a numeric variable}
\usage{
check_range(df, var)
}
\arguments{
\item{df}{the data frame to check}

\item{var}{numeric variable}
}
\description{
Returns error message when the value is out of range
}
\examples{
\dontrun{check_range(convert_choiceDF(choice_df), 'other_support_1')}
}
