% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_vars}
\alias{check_vars}
\title{Check data integrity}
\usage{
check_vars(num_df, choice_df, vars)
}
\arguments{
\item{num_df}{Numeric data frame}

\item{choice_df}{Choice data frame}

\item{vars}{A vector of variables to check}
}
\value{
TRUE, or global variable `.last.error`, if exception
}
\description{
Check data integrity by comparing two datasets.
}
