% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convert_choiceDF}
\alias{convert_choiceDF}
\title{Convert choice Dataframe}
\usage{
convert_choiceDF(df, var_names)
}
\arguments{
\item{var_names}{}
}
\description{
Convert choice Dataframe to remove text.
}
