% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convert_names}
\alias{convert_names}
\title{Simple conversion of names to make them consistent}
\usage{
convert_names(df, extra_map = NULL)
}
\arguments{
\item{df}{Input data frame}

\item{extra_map}{A named list of `new_var = old_var` for extra variable names mapping}
}
\value{
Names converted
}
\description{
Simple conversion of names to make them consistent
}
