% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convert_spss}
\alias{convert_spss}
\title{Convert SAV, SPSS to normal CSV}
\usage{
convert_spss(spss_path)
}
\arguments{
\item{df}{Path to spss file}
}
\value{
Converted data frame
}
\description{
Convert SAV, SPSS to normal CSV
}
