% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convert_start_end}
\alias{convert_start_end}
\title{Convert StartDate and EndDate to POSIXct}
\usage{
convert_start_end(df, date_format = NULL)
}
\arguments{
\item{df}{Data frame to parse}

\item{date_format}{Date format to parse}
}
\value{
POSIXct type columns
}
\description{
Convert StartDate and EndDate to POSIXct
}
