% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convert_time}
\alias{convert_time}
\title{Convert a vector to POSIXct}
\usage{
convert_time(x, date_format = NULL)
}
\arguments{
\item{x}{A vector as supported by lubridate::ymd_hm}

\item{date_format}{Date format to parse}
}
\value{
POSIXct type
}
\description{
Convert a vector to POSIXct
}
