% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_country_code}
\alias{get_country_code}
\alias{create_country_and_site}
\title{Get country code}
\usage{
get_country_code(country)

create_country_and_site(df, code, site)
}
\arguments{
\item{country}{character, Name of the country}

\item{df}{A Dataframe}

\item{code}{country code to include}

\item{site}{site name for the dataset}
}
\value{
The country code of the Country name

A dataframe with country code and site
        name added as 2 new columns
}
\description{
Get country code of the country name.
}
