% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{A data frame of the country and country code mapping}
\format{A data frame of Countries and the corresponding country codes}
\usage{
country_codes
}
\description{
A data frame of the country and country code mapping
}
\keyword{datasets}
