% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{create_country_and_site}
\alias{create_country_and_site}
\title{Add country code and site name for a data set}
\usage{
create_country_and_site(df, code, site)
}
\arguments{
\item{df}{A Dataframe}

\item{code}{country code to include}

\item{site}{site name for the dataset}
}
\value{
A dataframe with country code and site
        name added as 2 new columns
}
\description{
Add country code and site name for a data set
}
