% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_src_folders}
\alias{create_src_folders}
\title{Create a list of folders for each site}
\usage{
create_src_folders(folder_list)
}
\arguments{
\item{folder_list}{}
}
\description{
Create a list of folders for each site
}
