% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{diagnose}
\alias{diagnose}
\title{Diagnose potential error in the check}
\usage{
diagnose(rows, cols = NULL, df1, df2, last.error = TRUE)
}
\arguments{
\item{last.error}{}
}
\description{
Diagnose potential error from the check by comparing the
   choice and numeric data frames with selecting rows.
}
