% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_exc_vars}
\alias{get_exc_vars}
\title{Get a list of excluded variables from the merged df}
\usage{
get_exc_vars(non_num_vars_path, other_vars)
}
\arguments{
\item{non_num_vars_path}{Path to all non-numeric variables for different sites}

\item{other_vars}{other non_numeric variables that are included in the merged dataframe.
There variables will not be in the list of excluded variables.}
}
\value{
A list containing common  and sit-spec non-numeric variables
}
\description{
Get a list of excluded variables from the merged df
}
