% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_num_vars}
\alias{get_num_vars}
\title{Get numeric variables}
\usage{
get_num_vars(vars, pattern = NULL)
}
\arguments{
\item{pattern}{}
}
\description{
Get numeric variables by
  remove survey general variables and text variables.
}
