% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{get_problem_cols}
\alias{get_problem_cols}
\title{Identify columns that are not consistent}
\usage{
get_problem_cols(df)
}
\arguments{
\item{df}{DataFrame from \code{`ucom::diagnose()`}}
}
\value{
DataFrame with inconsistent columns
}
\description{
Identify columns that are not consistent
}
