% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rel_path}
\alias{get_rel_path}
\title{Relative path to UCOM root}
\usage{
get_rel_path(path)
}
\arguments{
\item{path}{a file system path inside UCOM project}
}
\value{
A relative path to the UCOM project root
}
\description{
Relative path to UCOM root
}
