% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_summary}
\alias{get_summary}
\title{Get summary of a cleaned data set}
\usage{
get_summary(df)
}
\arguments{
\item{df}{The cleaned data set}
}
\value{
A list of summary about the data set.
}
\description{
Get summary of a cleaned data set
}
\details{
It contains : name, country, status (1),
     dimensions, and proportion of missing values
}
