% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{org_gen_vars}
\alias{org_gen_vars}
\title{Survey general variables original - non-numeric}
\format{A vector of the survey original general variables}
\usage{
org_gen_vars
}
\description{
Survey general variables original - non-numeric
}
\keyword{datasets}
