% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_raw_data}
\alias{get_raw_data}
\alias{read_cleaned_data}
\alias{write_vars_rds}
\alias{write_results}
\title{Input raw data}
\usage{
get_raw_data(site, type, start_row = 3, file_name = NULL,
  sav = FALSE)

read_cleaned_data(df_path, num_vars_path, other_vars)

write_vars_rds(path, num_vars, non_num_vars)

write_results(choice_df, all_vars, num_vars, country_code = NULL)
}
\arguments{
\item{site}{Data folder (site) name}

\item{type}{Numeric or Choice value data}

\item{start_row}{number of rows to skip meta-data (description on variables)}

\item{file_name}{If provided, will read in this file rather than figure out itself}

\item{sav}{If the data file is SPSS file. Default FALSE.}

\item{df_path}{Cleaned data path}

\item{num_vars_path}{Path to the related num_vars.json}

\item{other_vars}{Other variables to include except for numeric}

\item{path}{path to output num_vars.rds and non_num_vars.rds}

\item{num_vars}{a vector of numeric variables}

\item{non_num_vars}{a vector of non-numeric variables}

\item{choice_df}{Choice data frame or transformed one if necessary}

\item{all_vars}{a vector of all variables}

\item{country_code}{Default NULL, if provided, will use the provided one instead of
figuring out itself.}
}
\value{
A data frame

A data frame

`num_vars.rds` & `non_num_vars.rds` to the output path

A csv file with the site name: `cleaned_data.csv`
}
\description{
Input raw data

Read in cleaned data for merging

Write the result: variable .rds file

Write the result: data .csv file
}
