% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_cleaned_data}
\alias{read_cleaned_data}
\title{Read in cleaned data for merging}
\usage{
read_cleaned_data(df_path, num_vars_path, other_vars)
}
\arguments{
\item{df_path}{Cleaned data path}

\item{num_vars_path}{Path to the related num_vars.json}

\item{other_vars}{Other variables to include except for numeric}
}
\description{
Read in cleaned data for merging
}
