% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{rename_uni_vars}
\alias{rename_uni_vars}
\title{Rename relevant variables to `uni` & `uni_text`}
\usage{
rename_uni_vars(df, site)
}
\arguments{
\item{df}{A DataFrame}

\item{site}{Site to check if the mapping exist}
}
\value{
A DataFrame with variable, if any, renamed to `uni`& `uni_text`
}
\description{
Rename relevant variables to `uni` & `uni_text`
}
