% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{text_vars}
\alias{text_vars}
\title{Text variables - non-numeric}
\format{A vector of the survey text variables}
\usage{
text_vars
}
\description{
Text variables - non-numeric
}
\keyword{datasets}
