% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vars_map}
\alias{vars_map}
\title{Mapping for unusal variables names to normal variable names}
\format{A named character vector}
\usage{
vars_map
}
\description{
Mapping for unusal variables names to normal variable names
}
\keyword{datasets}
