% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vars_to_change}
\alias{vars_to_change}
\title{Mapping of variables to `uni` or `uni_textbox` for each site}
\format{A data frame with 3 variables
\describe{
  \item{site}{Site name}
  \item{var_to_uni}{Variable to map to `uni`}
  \item{var_to_uni_textbox}{Variable to map to `uni_textbox`}
}}
\usage{
vars_to_change
}
\description{
Mapping of variables to `uni` or `uni_textbox` for each site
}
\keyword{datasets}
