% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_results}
\alias{write_results}
\title{Write the result: data .csv file}
\usage{
write_results(choice_df, all_vars, num_vars, country_code = NULL,
  output_path = NULL)
}
\arguments{
\item{output_path}{}
}
\description{
Write the result: data .csv file
}
