% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_vars_rds}
\alias{write_vars_rds}
\title{Write the result: variable .rds file}
\usage{
write_vars_rds(path, all_vars, num_vars)
}
\arguments{
\item{path}{path to output num_vars.rds and non_num_vars.rds}

\item{all_vars}{a vector of all variables}

\item{num_vars}{a vector of numeric variables}
}
\description{
Write the result: variable .rds file
}
