% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convert_choiceDF}
\alias{convert_choiceDF}
\title{Convert choice Dataframe}
\usage{
convert_choiceDF(df, var_names)
}
\arguments{
\item{df}{Data frame}

\item{var_names}{A vecot of variable names to convert}
}
\value{
Converted choice data frame
}
\description{
Convert choice Dataframe to remove text.
}
