% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convert_choiceDF}
\alias{convert_choiceDF}
\title{Convert choice Dataframe}
\usage{
convert_choiceDF(df, var_names,
  pattern = "^(-)?[0-9]{1,5}(\\\\.[0-9]{1,4})?")
}
\arguments{
\item{df}{Data frame}

\item{var_names}{A vecot of variable names to convert}

\item{pattern}{Regex pattern to match the numbers.
Default: `^(-)?[0-9]{1,5}(\\.[0-9]{1,4})?`}
}
\value{
Converted choice data frame
}
\description{
Convert choice Dataframe to remove text.
}
