% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convert_names}
\alias{convert_names}
\title{Simple conversion of names to make them consistent}
\usage{
convert_names(df)
}
\arguments{
\item{df}{Input data frame}
}
\value{
Names converted
}
\description{
Simple conversion of names to make them consistent
}
