% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{diagnose}
\alias{diagnose}
\title{Diagnose potential error in the check}
\usage{
diagnose(rows, cols = NULL, df1, df2, last.error = TRUE)
}
\arguments{
\item{rows}{which row to check}

\item{cols}{which column to check}

\item{df1}{A data frame}

\item{df2}{A data frame}

\item{last.error}{if TRUE, the data frames will be retrived from `.last.error`}
}
\value{
A View to browse the data
}
\description{
Diagnose potential error from the check by combining to
   compare the choice and numeric data frames with selecting rows.
}
