% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gen_vars}
\alias{gen_vars}
\title{Survey general variables lower case - non-numeric}
\format{A vector of the survey general variables converted to lower case}
\usage{
gen_vars
}
\description{
Survey general variables lower case - non-numeric
}
\keyword{datasets}
