% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_results}
\alias{write_results}
\title{Write the result: data .csv file}
\usage{
write_results(choice_df, all_vars, num_vars, country_code = NULL)
}
\arguments{
\item{choice_df}{Choice data frame or transformed one if necessary}

\item{all_vars}{a vector of all variables}

\item{num_vars}{a vector of numeric variables}

\item{country_code}{Default NULL, if provided, will use the provided one instead of
figuring out itself.}
}
\description{
Write the result: data .csv file
}
