% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_vars_rds}
\alias{write_vars_rds}
\title{Write the result: variable .rds file}
\usage{
write_vars_rds(path, all_vars, num_vars)
}
\arguments{
\item{num_vars}{}
}
\description{
Write the result: variable .rds file
}
